[python] Pin FE_TONEAREST when folding floats to decimal strings
float_to_string renders IEEE bits with std::pow/std::round and an ostream
conversion, all of which honour the host FP rounding mode. The GOTO/symex
float model leaves that mode non-default (FE_UPWARD on the Linux CI), so a
folded f-string like f"{-1.5:07.2f}" rendered "-001.49" instead of
"-001.50", failing the fstring_zero_pad regression there while passing
locally. Pin FE_TONEAREST for the fold, matching CPython's round-half-to-even.

The %-format and str.format paths already carried private copies of this
guard; hoist one round_to_nearest_guard into a shared header and reuse it in
all three sites instead of adding a third copy.
